requirements.txt
# DFB Spesen Generator - Dependencies

python-dotenv==1.0.0

# Web Scraping
playwright==1.48.0
beautifulsoup4==4.12.3
lxml==5.3.0

# DOCX Generation (für später)
python-docx==1.1.2

# FastAPI & Web (für später)
# fastapi>=0.104.0
# uvicorn>=0.24.0